+++ app/dashboard/components/visitors/visitors-tab.tsx
@@ -1,7 +1,6 @@
-import { DataTable } from "@/components/ui/data-table";
-import React from "react";
-import { visitorsColumns } from "../../config/table-columns";
 import { IVisitorData } from "@/app/types";
+import { DataTable } from "@/components/ui/data-table";
+import { visitorsColumns } from "../../config/table-columns";
 
 export default function Visitors({ visitors }: { visitors: IVisitorData[] }) {
   return <DataTable columns={visitorsColumns} data={visitors} />;
